==> PopGenome_Analysis/analyses/01-tidy_mtdt.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = here::here())
```
## Tidy Metadata 

```{r}
library(tidyverse)
```

```{r}

#........................................................
# Read in NGS sample data
#........................................................
smpls <- readxl::read_excel("../scrape_pubseqs/vivid_seq_public_NGS.xlsx") %>%
  dplyr::select(c("iid", "acc", "country", "vividregion"))

drc <- data.frame(iid = c("D9U3K", "O6Y4K", "Q8J6O"),
                  acc = c("D9U3K", "O6Y4K", "Q8J6O"),
                  country = "CD",
                  vividregion = "AF")

smpls <- rbind.data.frame(smpls, drc)

#........................................................
# Read in Passed Sample Data
#........................................................
psdsmpls <- readRDS("data/derived_data/final_smpl_list.RDS")
psdsmpls <- tibble::tibble(acc = psdsmpls, passed = "Y")


#........................................................
# Read in Clonal Data
#........................................................
hapuid <- readRDS("data/derived_data/mtdna_uniquehap_list.RDS")
hapuid <- purrr::map(hapuid, "unihapcounts") %>% 
  dplyr::bind_rows() %>% 
  dplyr::select(c("smpls")) %>% 
  dplyr::rename(acc = smpls) %>% 
  dplyr::mutate(hapuid = "Y")

#........................................................
# Bring it together
#........................................................
mtdt <- dplyr::left_join(smpls, psdsmpls, by = "acc") %>% 
  dplyr::left_join(., hapuid, by = "acc") %>% 
  dplyr::rename(smpls = acc) 

#......................
# add back in hapuid indicator for lab strains
#......................
mtdt <- mtdt %>% 
  dplyr::mutate(hapuid = ifelse(vividregion == "Lab", "Y", hapuid))

```
### Summary of Data
Overall, there are `r nrow(hapuid)` unique haplotypes in the data. Among the original `r nrow(mtdt)` samples considered (note, 3 are DRC samples), `r sum(mtdt$passed == "Y", na.rm=T)` passed QC thresholds. 

```{r}
#---------------------------------------------------- 
# Get Country Centroids
#---------------------------------------------------- 
# https://gis.stackexchange.com/questions/71921/list-of-central-coordinates-centroid-for-all-countries
wmap <- rworldmap::getMap(resolution="high")
centroids <- rgeos::gCentroid(wmap, byid=TRUE)
centroids <- as.data.frame(centroids)
centroids$country <- rownames(centroids)

# liftover country codes
centroids$country <- countrycode::countrycode(sourcevar = centroids$country, 
                                              origin = "country.name",
                                              destination = "iso2c")


#........................................................
# Bring it together
#........................................................
mtdt <- dplyr::left_join(mtdt, centroids, by = "country") 

```

### Perform Geographic K-means for Regionality
I broke the countries into various "political" regions (`vivid_regions`)... However, space doesn't follow these arbitrary borders... Will use K-means to partition data for statistics that need discrete classes (e.g. pairwise-$F_st$, Hudson's $S_nn$, etc.) for those samples that passed QC thresholds.

```{r}
coords <- mtdt %>% 
  dplyr::filter(smpls %in% psdsmpls$acc) %>% # passed samples only
  dplyr::filter(vividregion != "NHA" & country != "CD" 
                & vividregion != "Lab"  & country != "ES") %>% # exclude NHA & DRC & Ebro which will get their own cluster
  dplyr::select(c("country", "x", "y")) %>% 
  dplyr::filter(!duplicated(.)) 

#.............
# K-means Clustering of Globe
#.............
keda <- data.frame(k = seq(1, 16, by = 1))
keda$kmeans <- map(keda$k, function(k){return(kmeans(x = coords[,2:3], centers = k))})
keda$wss <- map(keda$kmeans, "withinss")
keda$totalwss <- map(keda$wss, function(x){return(sum(x))})

```

**First**, we need to find the inflection point to determine the optimal number of K-clusters (avoiding overfitting).
```{r, results='asis'}

keda.df <- keda %>% 
  dplyr::select(c("k", "totalwss")) %>% 
  tidyr::unnest(cols = totalwss)

kesteda.plotObj <- keda.df %>% 
  tibble::as_tibble(.) %>% 
  ggplot() +
  geom_line(aes(x=k, y=totalwss)) + 
  geom_point(aes(x=k, y=totalwss)) +
  geom_vline(xintercept = 4, color = "red", linetype = 2, alpha = 0.8) +
  theme_minimal() + 
  ylab("Total Within-Cluster Sum of Squares") +
  xlab("K")
plot(kesteda.plotObj)


```
**Four K groups looks reasonable to me**. Will also make Non-human apes and DRC it's own group, which was not considered above.

```{r}
set.seed(48)
k <- kmeans(coords[,2:3], 4)
globepart <- cbind.data.frame(coords, k = k$cluster) %>% 
  dplyr::select(c("country", "k"))

mtdt <- dplyr::left_join(mtdt, globepart, by = "country") %>%
  dplyr::filter(smpls %in% psdsmpls$acc) %>% # passed samples only
  dplyr::mutate(k = ifelse(is.na(k) & vividregion == "NHA", 5, k),
                k = ifelse(country == "CD", 6, k),
                k = ifelse(country == "ES", 7, k),
                k = factor(k, levels = c(1:7), labels = c("Africa", "Oceania", "Asia", "Americas", "NHA", "DRC", "Ebro1944")))


#---------------------------------------------------- 
# Get World Background
#---------------------------------------------------- 
world <- rnaturalearth::ne_countries(scale = "medium", returnclass = "sf") 


pvisolatpartitions <- ggplot() + 
  geom_sf(data = world, color = "#f0f0f0") +
  geom_point(data = mtdt, aes(x = x, y = y, color = factor(k))) +
  scale_colour_brewer("Region", palette = "Set1") +
  ggthemes::theme_map() +
  theme(legend.position = "bottom",
        legend.title = element_text(face = "bold"))

```

```{r, results='asis'}
plot(pvisolatpartitions)
```

```{r}
dir.create("figures/", recursive = T)
jpeg("figures/Pvglobal_partitions.jpg", width = 11, height = 8, units = "in", res = 500)
cowplot::plot_grid(kesteda.plotObj, pvisolatpartitions, align = "v", rel_widths = c(1, 2))
graphics.off()
```

```{r}
# housekeeping
dir.create("data/derived_data/", recursive = T)
saveRDS(mtdt, file = "data/derived_data/mtdt.RDS")

```
